+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ## [In Development] - Unreleased
 
 
+## [1.13.2] - 2022-05-18
+
+### Fixed
+
+- EveCharacter has no userprofile (#62)
+
+### Changed
+
+- HTML for ajax call result moved to its own template
+
+
 ## [1.13.1] - 2022-03-06
 
 ### Fixed